refactor: move business logic to service file
