+++ src/app/pipelines/overview/page.tsx
@@ -1,9 +1,7 @@
 import { useWorkspaceStore } from "@/state/stores/workspace-store";
 import { usePipelines } from "@/data/pipelines/all-pipelines-query";
-import { ChangeEvent, useEffect, useState } from "react";
-import { PipelineQueryParams } from "@/types/pipelines";
+import { ChangeEvent, useEffect } from "react";
 import Pagination from "@/components/pagination/Pagination";
-import { useSearchParams } from "react-router-dom";
 import BasePage from "@/components/common/BasePage";
 import { DataTable } from "@/components/table/DataTable";
 import { columns } from "./TableDef";
@@ -13,22 +11,16 @@ import { ReactComponent as Magnifying } from "@/assets/magnifying.svg";
 import { produce } from "immer";
 import debounce from "lodash.debounce";
 import { useDebouncedSearch } from "@/lib/hooks/debounce";
+import { usePagination } from "@/lib/hooks/pagination";
+import { usePipelinesOverviewService } from "./service";
 
 function Pipelines() {
-	const DEFAULT_PAGE = "1";
-	const DEFAULT_PAGE_SIZE = "10";
 	const activeWorkspace = useWorkspaceStore((state) => state.activeWorkspace);
 
-	const [searchParams, setSearchParams] = useSearchParams();
+	const { page, size } = usePagination();
 
-	const page = searchParams.get("page");
-	const size = searchParams.get("size");
-
-	const [queryParams, setQueryParams] = useState<PipelineQueryParams>({
-		page: parseInt(page || DEFAULT_PAGE),
-		size: parseInt(size || DEFAULT_PAGE_SIZE),
-		name: undefined
-	});
+	const { setQueryParams, queryParams, pageChangeHandler, pageSizeChangeHandler } =
+		usePipelinesOverviewService({ page, size });
 
 	const { data, isLoading, isError, isSuccess } = usePipelines({
 		workspace: activeWorkspace,
@@ -36,44 +28,10 @@ function Pipelines() {
 	});
 
 	useEffect(() => {
-		if (!searchParams.has("page")) {
-			searchParams.set("page", DEFAULT_PAGE);
-		}
-		if (!searchParams.has("size")) {
-			searchParams.set("size", DEFAULT_PAGE_SIZE);
-		}
-
-		setSearchParams(searchParams);
 		return () => {
 			debouncedSearch.cancel();
 		};
-	}, [searchParams, setSearchParams]);
-
-	function pageChangeHandler(page: number) {
-		setQueryParams((prevParams) => ({
-			...prevParams,
-			page
-		}));
-		setSearchParams((existing) => {
-			const newSearchParams = new URLSearchParams(existing.toString());
-			newSearchParams.set("page", page.toString());
-			return newSearchParams;
-		});
-	}
-
-	function pageSizeChangeHandler(size: number) {
-		setQueryParams((prevParams) => ({
-			...prevParams,
-			size,
-			page: 1
-		}));
-		setSearchParams((existing) => {
-			const newSearchParams = new URLSearchParams(existing.toString());
-			newSearchParams.set("size", size.toString());
-			newSearchParams.set("page", "1");
-			return newSearchParams;
-		});
-	}
+	}, []);
 
 	const debouncedSearch = useDebouncedSearch(
 		debounce((value: string) => {
